added code for basic product api

+++ src/main/java/com/technath/einventory/dao/repository/ProductRepository.java
@@ -4,8 +4,10 @@ import com.technath.einventory.dao.model.Product;
 import com.technath.einventory.dao.model.Supplier;
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigInteger;
+
 public interface ProductRepository extends CrudRepository<Product, Long> {
 
-    public Product findById(Long productId);
+    public Product findById(BigInteger productId);
 
 }
